feat: add a working implementation of tic-tac-toe
First pass of a working game API for Tic-tac-toe.

Endpoints:

- `POST /api/games`

Send a POST request to `/api/games` and a new Tic-tac-toe game will be created.
No request body is required or supported. The response body will be in JSON format.

Example response:

`201 Created`

```json
{
	"board": {
		"cells": ["", "", "", "", "", "", "", "", ""],
		"winningIndexTrio": null
	},
	"hasEnded": false,
	"players": [{
		"isTheirTurn": true,
		"isWinner": null,
		"name": "Player O",
		"uuid": "11111111-1111-1111-1111-111111111111"
	}, {
		"isTheirTurn": false,
		"isWinner": null,
		"name": "Player X",
		"uuid": "22222222-2222-2222-2222-222222222222"
	}],
	"uuid": "00000000-0000-0000-0000-000000000000"
}
```

Make note of the players’ UUIDs from the POST response: you’ll need these to validate a turn later,
because player UUIDs are omitted from any subsequent responses. This mechanism is intended to provide
rudementary authentication of turns, while keeping the API simple. Think of a player UUID as an
API key for that player. The downside is that whoever creates the game must be trusted to know
both players’ UUIDs!

- `GET /api/games/:uuid`

Send a GET request to `/api/games/:gameUuid` to receive the current state of your Tic-tac-toe game.

Example response:

`200 OK`

```json
{
	"board": {
		"cells": ["", "", "", "", "", "", "", "", ""],
		"winningIndexTrio": null
	},
	"hasEnded": false,
	"players": [{
		"isTheirTurn": true,
		"isWinner": null,
		"name": "Player O"
	}, {
		"isTheirTurn": false,
		"isWinner": null,
		"name": "Player X"
	}],
	"uuid": "00000000-0000-0000-0000-000000000000"
}
```

- `POST /api/games/:gameUuid/turn`

Calls to `/api/games/:gameUuid/turn` will change the state of the game board by taking the requested player’s turn in the game. If a successful response is returned, it will then be the other player’s turn.

Example request:

```json
{
	"cellToClaim": 3,
	"playerUuid": "11111111-1111-1111-1111-111111111111"
}
```

`cellToClaim` should be an integer from 0 through 8, where cells `0`, `1`, and `2` are the first row, `3`, `4`, and `5` are the second, and `6`, `7`, and `8` represent cells from the third row.

`playerUuid` should be the UUID of the player taking a turn. Player UUIDs were given on creation of the game.

Example response:

`200 OK`

```json
{
	"board": {
		"cells": ["", "", "", "O", "", "", "", "", ""],
		"winningIndexTrio": null
	},
	"hasEnded": false,
	"players": [{
		"isTheirTurn": false,
		"isWinner": null,
		"name": "Player O"
	}, {
		"isTheirTurn": true,
		"isWinner": null,
		"name": "Player X"
	}],
	"uuid": "00000000-0000-0000-0000-000000000000"
}
```

Technical constraints:

This API stores games in memory rather than in a dedicated database, in order to keep the architecture simple. It was designed to run on a free [Heroku](https://www.heroku.com/home) dyno with the assumption that when hosted, the service running this code will be periodically restarted. If you’re hosting this API somewhere other than Heroku and that service doesn’t periodically restart your app, be aware that your games array will grow until NodeJS runs out of memory.