feat: implement optional header for identifying a player
This header is useful for inferring the state of a player, for example: if you want to
serve an HTML page to the player, in order to infer if it’s their turn in the game, you need to know
which of the players in the players array is actually them. In other words, sending the `Player-UUID` header lets you identify a player within the response’s players array.