refactor: change id length for nicer urls
Problem: Game and player IDs make a full turn URL really long and
unreadable. These URLs are meant to be easy to share between players so it seems important to keep them short.

Solution:
We don’t really need 36 character UUIDs for game and player IDs because
we’re prettty unlikely to ever have an ID clash with like, at most, 20 or so active
games at a time! These games are meant to be short-lived, so it should be
fine to shorten the IDs. This commit changes them to be alphnumeric 5
character strings.

BREAKING CHANGE: Any mention of `uuid` or `playerUuid` in responses will
now be `id` and `playerId`, respectively. ID lengths are also now 5
alphanumeric characters instead of the 36 character hex (plus dash)
format of v4 UUIDs.