feature: enable two players to play subsequent games
When a game ends, a new game is created automatically and its data is shared in
any subsequent responses, with the same mechanism in place that omits player IDs
unless you specify one with a `Player-ID` header in the request. This approach
of perpetually creating subsequent games makes it possible for two players to
continue playing each other across multiple games, without having to share new
URLs between each other.