refactor: rename a property in the player model for clarity
Having a player property called `isTheirTurn`, in hindsight, wasn’t a great choice: using the singular third-person for a player makes it unclear whether a property refers to that player or their opponent. For example, if `player.isTheirTurn` is `true`, does that mean it’s that player’s turn, or is it their opponent’s turn? By renaming the property to `isTurn`, the problem is avoided.

BREAKING CHANGE: Any references to `isTheirTurn` will need to be changed to `isTurn`.