refactor: change mechanism for including player id in response
I’m playing around with a low-cost CDN (bunny.net) that I’m using for reverse-proxying and caching of this game. That’s quite fun but it turns out, the CDN doesn’t support varying the cache on a custom header (in this case, 'Player-ID'), so I’d have some cache clashing issues soon when I start introducing more caching to the game. Rabbit does support varying on query string values though, so this commit is the workaround.

BREAKING CHANGE: For anyone(?) relying on the `Player-ID` header to trigger the inclusion of playerIds in responses: You’ll need to change requests to include a query parameter `playerId`, in replacement of the `Player-ID` header. The query param value should be the same as old header’s value.